requirements.txt
requests
tenacity
schedule
python-dotenv